+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * #current_connection=
   * #current_clock=
   * #slave_consistent?
+  * ActiveRecord::Base.on_commit and ActiveRecord::Base.on_rollback
 * Fix 1.8.7 compliance
 
 # 0.2.0 (April 2, 2012)